- 모니터링 서버와 통신 관련 설정 추가 - 모니터링 서버와 통신 기능 추가 - 자기가 있는 위치로 이동하는 경우 수정 - 근처 적 탐색하는 함수 버그 수정

+++ GameServer/FieldObjectManager.cpp
@@ -100,7 +100,7 @@ void psh::FieldObjectManager::SpawnItem(psh::FVector loc, char type)
 {
     auto id = NextObjectId();
 
-    auto [item,success] = _items.emplace(id, make_shared<Item>(id,*static_cast<ObjectManager*>(this),_owner, loc,20, type));
+    auto [item,success] = _items.emplace(id, make_shared<Item>(id,*static_cast<ObjectManager*>(this),_owner, loc,static_cast<float>( 20), type));
     if (!success)
     {
         return;
